tabular/requirements.txt
numpy
pandas
torch 
torchvision 
cython 
ninja
scikit_learn
scipy
joblib
tensorboardX
tqdm
matplotlib
pyyaml
einops